added circ ref check
added js imp of a circ ref check, but its not as fast as the C++ equiv...
still adding it for record of having it somewhere as ill use similiar strat for
trying it in C++ instead with similar technique (also this technique for other issues open)

this technique will not recurse into self for deep nesting.

also, current c++ circ check falsely identifies b={c:'d'}; a=[b,b]; as circular when its not
my js method would not falsely identify it.